+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 
 - Change field names for user creation [#498](https://github.com/openkfw/TruBudget/issues/498)
+- Hide project search button instead of disabling it [#497](https://github.com/openkfw/TruBudget/issues/497)
 
 <!-- ### Deprecated -->